move the current mdns discovery implementation to discovery_legacy
