+++ examples/chat-with-mdns/README.md
@@ -57,12 +57,12 @@ func handleStream(stream net.Stream) {
 
 3. **Find peers nearby using mdns**
 
-Start [mdns discovery](https://godoc.org/github.com/libp2p/go-libp2p/p2p/discovery#NewMdnsService) service in host.
+Start [mdns discovery](https://godoc.org/github.com/libp2p/go-libp2p/p2p/discovery_legacy#NewMdnsService) service in host.
 
 ```go
 ser, err := discovery.NewMdnsService(ctx, peerhost, time.Hour, rendezvous)
 ```
-register [Notifee interface](https://godoc.org/github.com/libp2p/go-libp2p/p2p/discovery#Notifee) with service so that we get notified about peer discovery
+register [Notifee interface](https://godoc.org/github.com/libp2p/go-libp2p/p2p/discovery_legacy#Notifee) with service so that we get notified about peer discovery
 
 ```go
 	n := &discoveryNotifee{}
