+++ examples/chat-with-mdns/mdns.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/libp2p/go-libp2p/p2p/discovery"
+	discovery "github.com/libp2p/go-libp2p/p2p/discovery_legacy"
 )
 
 type discoveryNotifee struct {
